+++ proofs/sql.php
@@ -16,16 +16,13 @@ use Formal\ORM\{
 };
 use Formal\AccessLayer\Query;
 use Innmind\OperatingSystem\Factory;
-use Innmind\TimeContinuum\PointInTime;
+use Innmind\Time\Point;
 use Innmind\Filesystem\{
-    Adapter\InMemory,
+    Adapter,
     File,
 };
 use Innmind\Url\Url;
-use Innmind\Immutable\{
-    Sequence,
-    Either,
-};
+use Innmind\Immutable\Sequence;
 use Innmind\BlackBox\Set;
 
 return static function() {
@@ -38,29 +35,30 @@ return static function() {
         'SQL migrations',
         given(
             Set\MutuallyExclusive::of(
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
             ),
         ),
         static function($assert, $names) use ($os, $dsn) {
             [$a, $b, $c, $d] = $names;
 
             // setup
-            $os
+            $_ = $os
                 ->remote()
-                ->sql($dsn)(
-                    Query\SQL::of('drop table if exists `test`'),
+                ->sql($dsn)
+                ->unwrap()(
+                    Query::of('drop table if exists `test`'),
                 );
 
             $migrations = SQL::of(
                 $storage = Manager::filesystem(
-                    InMemory::emulateFilesystem(),
+                    Adapter::inMemory(),
                     Aggregates::of(
                         Types::of(
                             Support::class(
-                                PointInTime::class,
+                                Point::class,
                                 PointInTimeType::new($os->clock()),
                             ),
                         ),
@@ -73,25 +71,25 @@ return static function() {
             [$successfully, $versions] = $migrations(Sequence::of(
                 Migration::of(
                     $a,
-                    Query\SQL::of('create table `test` (`value` int not null)'),
+                    Query::of('create table `test` (`value` int not null)'),
                 ),
                 Migration::of(
                     $b,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('insert into `test` values (1)'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('insert into `test` values (1)'),
+                    Query::of('commit'),
                 ),
                 Migration::of(
                     $c,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('insert into `test` values (3)'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('insert into `test` values (3)'),
+                    Query::of('commit'),
                 ),
                 Migration::of(
                     $d,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('delete from `test` where `value` > 2'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('delete from `test` where `value` > 2'),
+                    Query::of('commit'),
                 ),
             ))->match(
                 static fn($versions) => [true, $versions],
@@ -99,7 +97,7 @@ return static function() {
             );
 
             $assert->true($successfully);
-            $assert->count(4, $versions);
+            $assert->same(4, $versions->size());
             $assert->same(
                 [$a, $b, $c, $d],
                 $versions
@@ -135,8 +133,9 @@ return static function() {
                 [['value' => 1]],
                 $os
                     ->remote()
-                    ->sql($dsn)(
-                        Query\SQL::of('select * from `test`'),
+                    ->sql($dsn)
+                    ->unwrap()(
+                        Query::of('select * from `test`'),
                     )
                     ->map(static fn($row) => $row->toArray())
                     ->toList(),
@@ -148,29 +147,30 @@ return static function() {
         'SQL partial migrations',
         given(
             Set\MutuallyExclusive::of(
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
             ),
         ),
         static function($assert, $names) use ($os, $dsn) {
             [$a, $b, $c, $d] = $names;
 
             // setup
-            $os
+            $_ = $os
                 ->remote()
-                ->sql($dsn)(
-                    Query\SQL::of('drop table if exists `test`'),
+                ->sql($dsn)
+                ->unwrap()(
+                    Query::of('drop table if exists `test`'),
                 );
 
             $migrations = SQL::of(
                 $storage = Manager::filesystem(
-                    InMemory::emulateFilesystem(),
+                    Adapter::inMemory(),
                     Aggregates::of(
                         Types::of(
                             Support::class(
-                                PointInTime::class,
+                                Point::class,
                                 PointInTimeType::new($os->clock()),
                             ),
                         ),
@@ -180,38 +180,35 @@ return static function() {
                 $dsn,
             );
 
-            $storage->transactional(
-                static function() use ($storage, $os, $d) {
-                    $storage
-                        ->repository(Version::class)
-                        ->put(Version::new($d, $os->clock()));
-
-                    return Either::right(null);
-                },
+            $_ = $storage->transactional(
+                static fn() => $storage
+                    ->repository(Version::class)
+                    ->put(Version::new($d, $os->clock()))
+                    ->either(),
             );
 
             [$successfully, $versions] = $migrations(Sequence::of(
                 Migration::of(
                     $a,
-                    Query\SQL::of('create table `test` (`value` int not null)'),
+                    Query::of('create table `test` (`value` int not null)'),
                 ),
                 Migration::of(
                     $b,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('insert into `test` values (1)'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('insert into `test` values (1)'),
+                    Query::of('commit'),
                 ),
                 Migration::of(
                     $c,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('insert into `test` values (3)'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('insert into `test` values (3)'),
+                    Query::of('commit'),
                 ),
                 Migration::of(
                     $d,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('delete from `test` where `value` > 2'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('delete from `test` where `value` > 2'),
+                    Query::of('commit'),
                 ),
             ))->match(
                 static fn($versions) => [true, $versions],
@@ -219,7 +216,7 @@ return static function() {
             );
 
             $assert->true($successfully);
-            $assert->count(3, $versions);
+            $assert->same(3, $versions->size());
             $assert->same(
                 [$a, $b, $c],
                 $versions
@@ -255,8 +252,9 @@ return static function() {
                 [['value' => 1], ['value' => 3]],
                 $os
                     ->remote()
-                    ->sql($dsn)(
-                        Query\SQL::of('select * from `test`'),
+                    ->sql($dsn)
+                    ->unwrap()(
+                        Query::of('select * from `test`'),
                     )
                     ->map(static fn($row) => $row->toArray())
                     ->toList(),
@@ -268,10 +266,10 @@ return static function() {
         'SQL migrations from raw files',
         given(
             Set\MutuallyExclusive::of(
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
             ),
         ),
         static function($assert, $names) use ($os, $dsn) {
@@ -279,12 +277,13 @@ return static function() {
             [$a, $b, $c, $d] = $names;
 
             // setup
-            $os
+            $_ = $os
                 ->remote()
-                ->sql($dsn)(
-                    Query\SQL::of('drop table if exists `test`'),
+                ->sql($dsn)
+                ->unwrap()(
+                    Query::of('drop table if exists `test`'),
                 );
-            $filesystem = InMemory::emulateFilesystem();
+            $filesystem = Adapter::inMemory();
             $sql = [
                 ['create table `test` (`value` int not null)'],
                 [
@@ -309,19 +308,19 @@ return static function() {
             ];
 
             foreach ($names as $i => $name) {
-                $filesystem->add(File::named(
+                $_ = $filesystem->add(File::named(
                     $name,
                     File\Content::ofString(\implode("\n", $sql[$i])),
-                ));
+                ))->unwrap();
             }
 
             $migrations = SQL::of(
                 $storage = Manager::filesystem(
-                    InMemory::emulateFilesystem(),
+                    Adapter::inMemory(),
                     Aggregates::of(
                         Types::of(
                             Support::class(
-                                PointInTime::class,
+                                Point::class,
                                 PointInTimeType::new($os->clock()),
                             ),
                         ),
@@ -337,7 +336,7 @@ return static function() {
             );
 
             $assert->true($successfully);
-            $assert->count(4, $versions);
+            $assert->same(4, $versions->size());
             $assert->same(
                 [$a, $b, $c, $d],
                 $versions
@@ -373,8 +372,9 @@ return static function() {
                 [['value' => 1]],
                 $os
                     ->remote()
-                    ->sql($dsn)(
-                        Query\SQL::of('select * from `test`'),
+                    ->sql($dsn)
+                    ->unwrap()(
+                        Query::of('select * from `test`'),
                     )
                     ->map(static fn($row) => $row->toArray())
                     ->toList(),
@@ -386,28 +386,29 @@ return static function() {
         'SQL failing migrations',
         given(
             Set\MutuallyExclusive::of(
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
-                Set\Strings::madeOf(Set\Chars::alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
+                Set::strings()->madeOf(Set::strings()->chars()->alphanumerical())->atLeast(1),
             ),
         ),
         static function($assert, $names) use ($os, $dsn) {
             [$a, $b, $c] = $names;
 
             // setup
-            $os
+            $_ = $os
                 ->remote()
-                ->sql($dsn)(
-                    Query\SQL::of('drop table if exists `test`'),
+                ->sql($dsn)
+                ->unwrap()(
+                    Query::of('drop table if exists `test`'),
                 );
 
             $migrations = SQL::of(
                 $storage = Manager::filesystem(
-                    InMemory::emulateFilesystem(),
+                    Adapter::inMemory(),
                     Aggregates::of(
                         Types::of(
                             Support::class(
-                                PointInTime::class,
+                                Point::class,
                                 PointInTimeType::new($os->clock()),
                             ),
                         ),
@@ -420,17 +421,17 @@ return static function() {
             [$successfully, $versions, $error] = $migrations(Sequence::of(
                 Migration::of(
                     $a,
-                    Query\SQL::of('create table `test` (`value` int not null)'),
+                    Query::of('create table `test` (`value` int not null)'),
                 ),
                 Migration::of(
                     $b,
-                    Query\SQL::of('create table `test` (`value` int not null)'),
+                    Query::of('create table `test` (`value` int not null)'),
                 ),
                 Migration::of(
                     $c,
-                    Query\SQL::of('start transaction'),
-                    Query\SQL::of('insert into `test` values (3)'),
-                    Query\SQL::of('commit'),
+                    Query::of('start transaction'),
+                    Query::of('insert into `test` values (3)'),
+                    Query::of('commit'),
                 ),
             ))->match(
                 static fn($versions) => [true, $versions, null],
@@ -445,7 +446,7 @@ return static function() {
                 "Query 'create table `test` (`value` int not null)' failed with: [42S01] [1050] Table 'test' already exists",
                 $error->getMessage(),
             );
-            $assert->count(1, $versions);
+            $assert->same(1, $versions->size());
             $assert->same(
                 [$a],
                 $versions
@@ -473,8 +474,9 @@ return static function() {
                 [],
                 $os
                     ->remote()
-                    ->sql($dsn)(
-                        Query\SQL::of('select * from `test`'),
+                    ->sql($dsn)
+                    ->unwrap()(
+                        Query::of('select * from `test`'),
                     )
                     ->map(static fn($row) => $row->toArray())
                     ->toList(),