+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Requires PHP `8.4`
+- Requires `formal/orm:~6.0`
+- Requires `innmind/foundation:~2.1`
+
 ## 1.1.0 - 2025-02-28
 
 ### Added